Fix fractional values in numeric collection sorting
Multi-attribute SORT_NUMERIC comparisons truncated both values to integers,
so different fractional prices or scores became ties and were ordered by
secondary fields or input order. Compare floats to match PHP's native
numeric sort and Collection's single-attribute path.

Keep the existing comparison loop, sort directions, key preservation and
LazyCollection delegation. Add one shared regression for both collection
classes covering positive and negative fractions, numeric strings,
secondary-key ties and both primary sort directions.

Found while reconciling Laravel's casting update; the defect originated in
its multi-attribute sort-option implementation and remains in source pin
01d008c9b5f32cb7c5e50a9a22273113d810b2a2.

Upstream:
https://github.com/laravel/framework/pull/50269
https://github.com/laravel/framework/pull/58037

Validation: the new regression failed with the integer comparator and passes
with the float comparator. Affected collection, support, request and file
validation tests pass, as do formatting and both PHPStan configurations.

+++ src/collections/src/Collection.php
@@ -1572,7 +1572,7 @@ class Collection implements ArrayAccess, CanBeEscapedWhenCastToString, Enumerabl
                         }
                     } else {
                         $result = match ($options) {
-                            SORT_NUMERIC => (int) $values[0] <=> (int) $values[1],
+                            SORT_NUMERIC => (float) $values[0] <=> (float) $values[1],
                             SORT_STRING => strcmp((string) $values[0], (string) $values[1]),
                             SORT_NATURAL => strnatcmp((string) $values[0], (string) $values[1]),
                             SORT_LOCALE_STRING => strcoll((string) $values[0], (string) $values[1]),
